// src/modules/podcast/dto/search-podcast.dto.ts
export class SearchPostDto {
  user_id?: string;
  podcast_type?: string;
  search?: string;
  podcast_status?: string;
  podcast_language?: string;
  podcast_category?: string;
  not_image?: string;
  categories?: any;
  not_download?: string;
  other_status?: string;
  post_parent?: string;
  country?: string;
  from?: string;
  comment_number?: string;
  channel_id?: string;
}